requirements.txt
openai
fastapi
uvicorn
pymongo
python-dotenv
torch
marker-pdf
asyncio
langchain
python-multipart
langchain-openai 
langchain-community